fix: deprecated setBundles and test
