+++ plugins/standalone/package.json
@@ -49,6 +49,7 @@
     "@types/node": "^20.9.4",
     "@types/semver": "^7.5.8",
     "msw": "^2.7.0",
-    "semver": "^7.6.3"
+    "semver": "^7.6.3",
+    "vitest": "^3.0.6"
   }
 }
