--- plugins/standalone/src/types.ts
@@ -1,10 +1,0 @@
-export interface RouteConfig {
-  path: string;
-  headers?: Record<string, string>;
-}
-
-export interface Routes {
-  upsert: () => RouteConfig;
-  list: () => RouteConfig;
-  retrieve: (bundleId: string) => RouteConfig;
-}
